requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
sentence-transformers>=2.7.0
chromadb>=0.4.24
pymupdf>=1.24.0
python-docx>=1.1.0
